curiouscows/requirements.txt
Django==5.1.2
gensim==4.3.3
numpy==1.24.3
pandas==2.0.1
protobuf==5.28.3
python-dotenv==1.0.1
Requests==2.32.3
tqdm==4.66.5
gunicorn==20.1.0
google-generativeai==0.8.3